Python/requirements.txt
nltk-3.2.5
six-1.11.0
scikit-learn-0.19.1
sklearn-0.0
numpy-1.14.0
scipy-1.0.0
unidecode-1.0.22
tqdm-4.19.5

